+++ mlflow/prompt/registry_utils.py
@@ -1,5 +1,6 @@
 import functools
 import json
+import logging
 import re
 from textwrap import dedent
 from typing import Any
